price-reporter: Moved /external-events to /core/src/price_reporter.
