+++ Cargo.toml
@@ -4,7 +4,6 @@ members = [
 	"circuit-macros",
 	"core",
 	"crypto",
-    "external-events",
 	"integration-helpers"
 ]
