+++ core/Cargo.toml
@@ -7,19 +7,26 @@ edition = "2018"
 
 [dependencies]
 ark-serialize = "0.3"
-async-trait = "0.1.56"
+async-trait = "0.1.60"
 base64 = { version = "0.13" }
+bimap = "0.6.2"
 bus = { version = "2.3" }
 circuits = { path = "../circuits" }
+chrono = "0.4.23"
 clap = { version = "3.2.8", features = ["derive"] }
 crossbeam = { version = "0.8.1" }
+create2 = "0.0.2"
 crypto = { path = "../crypto" }
 curve25519-dalek = "2"
+dotenv = "0.15.0"
 ed25519-dalek = { version = "1.0.1" }
-futures = { version = "0.3.21" }
+futures = { version = "0.3.25" }
 futures-util = { version = "0.3" }
+hex = "0.3.1"
+hmac-sha256 = "1.1.6"
 hyper = { version = "0.14", features = ["http1", "http2", "server", "tcp"] }
 integration-helpers = { path = "../integration-helpers" }
+lazy_static = "1.4.0"
 libp2p = { version = "0.50", features = [
     "async-std",
     "dns",
@@ -45,13 +52,18 @@ portpicker = "0.1"
 rand = { version = "0.8.5", features = ["getrandom"] }
 rand_core = "0.5"
 rayon = { version = "1.5.3" }
+reqwest = "0.11.13"
+ring-channel = "0.11.0"
 serde = { version = "1.0.139", features = ["serde_derive"] }
 serde_json = "1.0"
+streaming-stats = "0.1.28"
 termion = "2.0"
 tokio = { version = "1", features = ["full"] }
 toml = { version = "0.5.9" }
 tracing = { version = "0.1", features = ["log"] }
 tokio-stream = { version = "0.1" }
 tokio-tungstenite = { version = "0.18" }
-tungstenite = { version = "0.18" }
+tungstenite = { version = "0.18.0", features = ["native-tls"] }
+url = "2.3.1"
 uuid = { version = "1.1.2", features = ["v4", "serde"] }
+web3 = "0.18.0"
