--- external-events/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "external-events"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-async-trait = "0.1.60"
-bimap = "0.6.2"
-chrono = "0.4.23"
-create2 = "0.0.2"
-dotenv = "0.15.0"
-futures = "0.3.25"
-hex = "0.3.1"
-hmac-sha256 = "1.1.6"
-lazy_static = "1.4.0"
-reqwest = "0.11.13"
-ring-channel = "0.11.0"
-serde = "1.0.139"
-serde_json = "1.0"
-streaming-stats = "0.1.28"
-tracing = { version = "0.1", features = ["log"] }
-tokio = { version = "1", features = ["full"] }
-tungstenite = { version = "0.18.0", features = ["native-tls"] }
-url = "2.3.1"
-web3 = "0.18.0"
